Move lights to interfaces
